requirements.txt
pandas==2.2.3
numpy==2.1.1
pyarrow==17.0.0
deltalake==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.2
streamlit==1.39.0
plotly==5.24.1
scikit-learn==1.5.2
transformers==4.45.2
torch==2.4.1
pytest==8.3.3
kafka-python==2.0.2